Change ints to long ints

+++ src/uhttp.c
@@ -344,8 +344,8 @@ static void log_data_line(HTTP_CLIENT_HANDLE_DATA* http_data, const char* text_l
     }
     else
     {
-        int authPos = authStart - text_line;
-        int proxyAuthPos = proxyAuthStart - text_line;
+        long int authPos = authStart - text_line;
+        long int proxyAuthPos = proxyAuthStart - text_line;
         if(authPos < proxyAuthPos)
         {
             LOG(AZ_LOG_TRACE, LOG_LINE, "%.*s \r\nAuthorization: *** %.*s \r\nProxy-Authorization: *** %.*s", (int)(authStart - text_line), text_line, (int)(proxyAuthStart - authEol), authEol, (int)(proxyAuthEol - text_line), proxyAuthEol);
